added new check roles users

+++ src/WarehouseBundle/Controller/ApiProductController.php
@@ -5,6 +5,7 @@ namespace WarehouseBundle\Controller;
 use FOS\RestBundle\Controller\FOSRestController;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use WarehouseBundle\Entity\Product;
 use WarehouseBundle\Form\ProductForm;
 use Symfony\Component\HttpFoundation\Response;
@@ -28,12 +29,16 @@ class ApiProductController extends FOSRestController
             'csrf_protection' => false,
         ]);
 
-        $view = $this->view($form->getData(), 200);
+        $view = $this->view($product, 200);
         return $this->handleView($view);
     }
 
     public function postProductAction(Request $request)
     {
+        $user = $this->get('security.token_storage')->getToken()->getUser();
+
+        $this->denyAccessUnlessGranted('create', $user);
+
         $form = $this->createForm(ProductForm::class, new Product(), [
             'csrf_protection' => false,
             'method' => $request->getMethod()
@@ -73,6 +78,10 @@ class ApiProductController extends FOSRestController
 
     public function putProductAction(Product $product, Request $request)
     {
+        $user = $this->get('security.token_storage')->getToken()->getUser();
+
+        $this->denyAccessUnlessGranted('edit', $user);
+
         $form = $this->createForm(ProductForm::class, $product, ['csrf_protection' => false, 'method' => $request->getMethod()]);
 
         return $this->processForm($request, $form);
@@ -82,6 +91,10 @@ class ApiProductController extends FOSRestController
 
     public function patchProductAction(Product $product, Request $request)
     {
+        $user = $this->get('security.token_storage')->getToken()->getUser();
+
+        $this->denyAccessUnlessGranted('edit', $user);
+
         $form = $this->createForm(ProductForm::class, $product, ['csrf_protection' => false, 'method' => $request->getMethod()]);
 
         return $this->processForm($request, $form);
@@ -89,6 +102,10 @@ class ApiProductController extends FOSRestController
 
     public function deleteProductAction(Product $product)
     {
+        $user = $this->get('security.token_storage')->getToken()->getUser();
+
+        $this->denyAccessUnlessGranted('delete', $user);
+
         if($product) {
             $em = $this->getDoctrine()->getManager();
             $em->remove($product);
